backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
PyGithub
anthropic
pydantic-settings